+++ main.ts
@@ -1,1 +1,4 @@
-console.log(`🚀`);
+import "./paths";
+import { config } from "@packages/config";
+
+console.log(`🚀 (${config.env})`);