+++ paths.ts
@@ -1,0 +1,4 @@
+// Monorepo paths resolution
+import moduleAlias from "module-alias";
+import { join } from "path";
+moduleAlias.addAliases({ "@packages": join(__dirname, "packages") });