+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 [0.2.0]
 - TypeScript configuration
 - ESLint / Prettier configuration
+- Monorepo paths resolution